+++ lib/exop/chain.ex
@@ -31,17 +31,26 @@ defmodule Exop.Chain do
   the chain execution interrupts and error result returned (as the chain (`CreateUser`) result).
   """
 
-  defmacro __using__(_opts) do
+  defmacro __using__(opts \\ []) do
     quote do
       import unquote(__MODULE__)
 
       Module.register_attribute(__MODULE__, :operations, accumulate: true)
+
+      @error_includes_operation_name unquote(opts)[:name_in_error] == true
 
       @before_compile unquote(__MODULE__)
     end
   end
 
   defmacro operation(operation, additional_params \\ []) do
+    quote bind_quoted: [operation: operation, additional_params: additional_params] do
+      {:module, operation} = Code.ensure_compiled(operation)
+      @operations %{operation: operation, additional_params: additional_params}
+    end
+  end
+
+  defmacro step(operation, additional_params \\ []) do
     quote bind_quoted: [operation: operation, additional_params: additional_params] do
       {:module, operation} = Code.ensure_compiled(operation)
       @operations %{operation: operation, additional_params: additional_params}
@@ -60,7 +69,7 @@ defmodule Exop.Chain do
       #  {:error, {:auth, :unknown_action}}   |
       #  {:error, {:auth, atom}}
 
-      @not_ok :not_ok
+      @not_ok :exop_not_ok
 
       @doc """
       Invokes all operations defined in a chain. Returns either a result of the last operation
@@ -73,9 +82,16 @@ defmodule Exop.Chain do
           ok_result = @operations |> Enum.reverse() |> invoke_operations({:ok, received_params})
           {:ok, ok_result}
         catch
-          {@not_ok, not_ok_result} -> not_ok_result
+          {@not_ok, not_ok_result, operation} ->
+            add_operation_name(@error_includes_operation_name, not_ok_result, operation)
+
+          {@not_ok, not_ok_result} ->
+            not_ok_result
         end
       end
+
+      defp add_operation_name(true, not_ok_result, operation), do: {operation, not_ok_result}
+      defp add_operation_name(_, not_ok_result, _), do: not_ok_result
 
       @spec invoke_operations([%{operation: atom(), additional_params: Keyword.t()}], any()) ::
               any()
@@ -89,11 +105,16 @@ defmodule Exop.Chain do
            ) do
         params = params |> merge_params(additional_params) |> resolve_params_values()
 
-        with {:ok, result} <- apply(operation, :run, [params]) do
-          result
-        else
-          not_ok ->
-            throw({@not_ok, not_ok})
+        case apply(operation, :run, [params]) do
+          result when is_tuple(result) and elem(result, 0) == :error ->
+            throw({@not_ok, result, operation})
+            @not_ok
+
+          {:ok, result} ->
+            result
+
+          result ->
+            throw({@not_ok, result})
             @not_ok
         end
       end
@@ -104,11 +125,16 @@ defmodule Exop.Chain do
            ) do
         params = params |> merge_params(additional_params) |> resolve_params_values()
 
-        with {:ok, _} = result <- apply(operation, :run, [params]) do
-          invoke_operations(tail, result)
-        else
-          not_ok ->
-            throw({@not_ok, not_ok})
+        case apply(operation, :run, [params]) do
+          result when is_tuple(result) and elem(result, 0) == :error ->
+            throw({@not_ok, result, operation})
+            @not_ok
+
+          {:ok, _} = result ->
+            invoke_operations(tail, result)
+
+          result ->
+            throw({@not_ok, result})
             @not_ok
         end
       end